pyproject.toml
[project]
name = "ALovelyBot"
version = "0.1.0"
description = "ALovelyBot"
readme = "README.md"
requires-python = ">=3.8, <4.0"
dependencies = [
    "nonebot-adapter-qq>=1.4.4",
    "nonebot2[aiohttp]>=2.2.1",
    "nonebot-plugin-fishing>=0.2.3",
]

[tool.nonebot]
adapters = [
    { name = "QQ", module_name = "nonebot.adapters.qq" }
]
plugins = ["nonebot_plugin_fishing"]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.pdm]
